Rename props to types, update docs
