+++ packages/experiments/src/components/StaticList/StaticList.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { css } from '../../Utilities';
 import { IObjectWithKey } from 'office-ui-fabric-react/lib/Selection';
-import { IStaticListProps } from './StaticList.Props';
+import { IStaticListProps } from './StaticList.types';
 
 import * as stylesImport from './StaticList.scss';
