+++ packages/experiments/src/components/StaticList/StaticList.types.ts
@@ -1,6 +1,7 @@
 import { IObjectWithKey } from 'office-ui-fabric-react/lib/Selection';
-import { IGenericListProps } from './List.Props';
+import { IGenericListProps } from './List.types';
 
 export interface IStaticListProps<TItem extends IObjectWithKey> extends IGenericListProps<TItem> {
+  /** Html tag to use for rendering the list, defaults to 'ul' */
   listTagName?: string;
 }
