+++ packages/experiments/src/components/StaticList/StaticListPage.tsx
@@ -17,7 +17,7 @@ export class StaticListPage extends React.Component<IComponentDemoPageProps, {}>
         propertiesTables={
           <PropertiesTableSet
             sources={ [
-              require<string>('!raw-loader!experiments/src/components/StaticList/StaticList.Props.ts')
+              require<string>('!raw-loader!experiments/src/components/StaticList/StaticList.types.ts')
             ] }
           />
         }
