+++ packages/experiments/src/components/VirtualizedList/VirtualizedList.types.ts
@@ -1,5 +1,5 @@
 import { IBaseProps } from '../../Utilities';
-import { IGenericListProps } from '../StaticList/List.Props';
+import { IGenericListProps } from '../StaticList/List.types';
 import { IObjectWithKey } from 'office-ui-fabric-react/lib/Selection';
 
 export interface IVirtualizedListProps<TItem extends IObjectWithKey> extends IGenericListProps<TItem>, IBaseProps {
@@ -12,5 +12,6 @@ export interface IVirtualizedListProps<TItem extends IObjectWithKey> extends IGe
   /** Number of items to draw before/after viewport height */
   itemOverdraw?: number;
 
+  /** Html tag to use for spacer elements in the list, defaults to 'div' */
   spacerItemTagName?: string;
 }
