+++ packages/experiments/src/components/VirtualizedList/index.ts
@@ -1,2 +1,2 @@
-export * from './VirtualizedList.Props';
+export * from './VirtualizedList.types';
 export * from './VirtualizedList';
